requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0
streamlit>=1.30.0
transformers>=4.40.0
datasets>=2.15.0
evaluate>=0.4.0
scikit-learn>=1.3.0
accelerate>=0.26.0
shapely>=2.0.0
